Updated MineSweep.
Added splash screen (created in Canvas) and visually reduced load time by splitting code.
